==> 2 - Sentiment - Tim.ipynb <==
# This notebook was fully populated in Tim Jenkins' AWS Account, which became locked due as we completed our image processing module. 

# The results of Tim's Sentiment efforts can be found in 'review_boulder_with_sentiment.csv'
