==> 3 - Image Detection - Tim.ipynb <==
# This file was fully populated in Tim Jenkins' AWS account, which became locked due to overuse. This notebook is incomplete as a result. 

# The results of Tim's work can be found in 'photos_boulder_with_labels.csv'

import glob
import boto3
import os
import pandas as pd

bucket='sagemaker-us-east-1-719086491758'
prefix = "final_images" # replace this with the path to your images
CSVIN='twitter_with_LDA.csv'
CSVOUT='twitter_with_LDA_and_image_data.csv'

df = pd.read_csv(CSVIN)
# Convert these numbers to categories
df['id'] = df['id'].astype('object')
print(df.shape)
df.head()

#limit number of rows for testing-if you don't want to process everything
df=df.head(2)
#df.dtypes
df2 = df
df

#get the rekognition client
client = boto3.client('rekognition')

#get the s3 client
s3 = boto3.resource('s3')
my_bucket = s3.Bucket(bucket)
#files = my_bucket.objects.filter(Prefix=prefix)

dfFaces = pd.DataFrame([])
i=0

for row in df.itertuples():
    file = f"{row.id}"
    extension = f"{row.url.split('.')[-1]}"
    print(f"Processing TweetID - {file}.{extension}")    
    if not pd.isnull(file):
        #print(i, row[1], file, f"{prefix}/{row[1]}.{file.split('.')[-1]}")
    
    
        if extension == 'jpg':
            # call rekognition for this next file
            response = client.detect_faces(
                Image={
                    'S3Object': {
                        'Bucket': bucket,
                        'Name': f"{prefix}/{file}.{extension}"
                    }
                },
                Attributes=[
                    'ALL',
                ]
            )

            # now add all of the facial features for every person found in the photo
            for fd in response["FaceDetails"]:    
                i=i+1
                dfFaces.loc[i,'TweetID'] = file
                dfFaces.loc[i,'PersonID'] = i
                dfFaces.loc[i,'AgeRange-Low'] = fd["AgeRange"]["Low"]
                dfFaces.loc[i,'AgeRange-High'] = fd["AgeRange"]["High"]
                dfFaces.loc[i,'Smile'] = fd["Smile"]["Value"]
                dfFaces.loc[i,'Gender'] = fd["Gender"]["Value"]
                dfFaces.loc[i,'Emotion'] = fd["Emotions"][0]["Type"] #get dominant emotion
                dfFaces.loc[i,'Emotion-Confidence'] = fd["Emotions"][0]["Confidence"] #get dominant emotion


dfFaces.describe

dfFaceSummary = pd.DataFrame([])

#summarize stats per file
#aggregate functions include min, max, mean, count, and more.
dfFaceSummary['Count-People'] = dfFaces.groupby('TweetID')['PersonID'].count()
dfFaceSummary['Avg-AgeRange-Low'] = dfFaces.groupby('TweetID')['AgeRange-Low'].min()
dfFaceSummary['Avg-AgeRange-High'] = dfFaces.groupby('TweetID')['AgeRange-High'].max()
dfFaceSummary['Count-Smile'] = dfFaces[dfFaces['Smile']==True].groupby('TweetID')['Smile'].count()
dfFaceSummary['Count-Smile'] = dfFaceSummary['Count-Smile'].fillna(0)
#****Fix this****
dfFaceSummary['Percent-Smile'] = dfFaceSummary['Count-Smile']/1
dfFaceSummary['Percent-Smile'] = dfFaceSummary['Percent-Smile'].fillna(0)

dfFaceSummary = dfFaceSummary.merge(dfFaces.groupby('TweetID')['Emotion'].value_counts().unstack().fillna(0), on='TweetID')
dfFaceSummary = dfFaceSummary.merge(dfFaces.groupby('TweetID')['Gender'].value_counts().unstack().fillna(0), on='TweetID')

#print out 
dfFaceSummary

#the tweet id is an object, so we need to convert it to the same data type as our id
dfFaceSummary.index = dfFaceSummary.index.astype(int)

#Now merge the frames together
df = df.merge(dfFaceSummary, left_on='id', right_on='TweetID')
df

#now save this out to a new csv
df.to_csv(CSVOUT, index=False)

dflabels = pd.DataFrame([])

for row in df2.itertuples():
    file = f"{row.id}"
    extension = f"{row.url.split('.')[-1]}"
    print(f"Processing TweetID - {file}.{extension}")    
    if not pd.isnull(file):
        #print(i, row[1], file, f"{prefix}/{row[1]}.{file.split('.')[-1]}")
    
    
        if extension == 'jpg':
            # call rekognition for this next file
            response = client.detect_labels(
            Image={
                    'S3Object': {
                        'Bucket': bucket,
                        'Name': f"{prefix}/{file}.{extension}"
                    }
            }, MaxLabels=10
        )

        #now add all of the labels found in this photo
        for lb in response["Labels"]:
            print (lb["Name"] + " " + str(lb["Confidence"]))
