==> 4 - Combine Dataframes.ipynb <==
# Read CSV inputs into Dataframes

import pandas as pd
import numpy as np

bucket = "bg-advancedanalytics-emba"
folder = "final-project"

business_na = {'attributes.WiFi':'Unknown', 'attributes.RestaurantsPriceRange2':'Unknown', 'attributes.Alcohol':'Unknown', 'attributes.RestaurantsTakeOut':'Unknown' }

business = pd.read_csv(f's3://{bucket}/{folder}/business_boulder.csv', low_memory=False).fillna('Unknown')
print("Business")
photo = pd.read_csv(f's3://{bucket}/{folder}/photos_boulder_with_labels.csv', low_memory=False)
print("Photo")
review = pd.read_csv(f's3://{bucket}/{folder}/review_boulder_with_sentiment_and_LDA.csv', low_memory=False)
print("Review")
user = pd.read_csv(f's3://{bucket}/{folder}/user_boulder.csv', low_memory=False)
print("User")

# Read Dataframes

business
# photo
# review
# user

# Group Photos Dataset
init_len = len(photo.columns)

for col in range(5, init_len):
#     print(col)
    colName = photo.columns[col] + " 1/0"
#     print(colName)
    cond = np.where(photo.iloc[:,col]>0,1,0)
#     print(cond)
    photo[colName] = cond
    
groups = list(photo)[init_len:len(photo.columns)]
photo = photo.groupby('business_id')[groups].agg('sum')
photo

print(list(photo))

# Create joined dataset

review_user = review.join(user.set_index('user_id'), on='user_id', how='left', rsuffix='_fromUser')
review_user_business = review_user.join(business.set_index('business_id'), on='business_id', how='left', rsuffix='_fromBusiness')
review_user_business_photo = review_user_business.join(photo, on='business_id', how='left', rsuffix='_fromPhoto').fillna(0)

review_user_business_photo.to_csv(f's3://{bucket}/{folder}/combined_dataset.csv', index=False)
review_user_business_photo.to_csv('combined_dataset.csv')
review_user_business_photo
